Adjust ConsumerScreen, add item generation logic to BasicConsumerEntity

+++ src/main/java/dev/shinyepo/resourcegenerator/configs/ConsumerConfig.java
@@ -1,15 +1,15 @@
 package dev.shinyepo.resourcegenerator.configs;
 
 public enum ConsumerConfig {
-    BASIC_CONSUMER(2L);
+    BASIC_CONSUMER(2);
 
-    private final Long produces;
+    private final int produces;
 
-    ConsumerConfig(long produces) {
+    ConsumerConfig(int produces) {
         this.produces = produces;
     }
 
-    public Long getProduces() {
+    public int getProduces() {
         return produces;
     }
 }
